requirements.txt
google-cloud-storage
pyserial
google-cloud-firestore